flask-microservice-tutorial4/requirements.txt
Flask==1.1.1
Flask-Injector==0.12.0
connexion==2.5.1
swagger-ui==0.1.2
swagger-ui-bundle==0.0.6
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
marshmallow==3.3.0
